+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   a more compact `{"id":"42.23",,...}`.
 * Omit the `etag` property and the `requestId` property if they are null rather than transmitting nulls.
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.13.0`.
+* Support `ChannelAddress.getName()` even when `Replicant.areNamesEnabled()` returns false as that behaviour
+  is required and was implemented in other places.
 
 ### [v6.32](https://github.com/replicant4j/replicant/tree/v6.32) (2019-02-24)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.31...v6.32)